+++ src/main/java/roomescape/application/dto/ReservationRequest.java
@@ -2,10 +2,10 @@ package roomescape.application.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
-import roomescape.domain.Member;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Theme;
+import roomescape.domain.member.Member;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationTime;
+import roomescape.domain.theme.Theme;
 
 public record ReservationRequest(
         @NotNull LocalDate date,
