+++ src/main/java/roomescape/application/dto/ReservationResponse.java
@@ -1,7 +1,7 @@
 package roomescape.application.dto;
 
 import java.time.LocalDate;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.Reservation;
 
 public record ReservationResponse(long id, MemberResponse member, LocalDate date,
                                   ReservationTimeResponse time,
