+++ src/main/java/roomescape/application/dto/ReservationTimeRequest.java
@@ -2,7 +2,7 @@ package roomescape.application.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservation.ReservationTime;
 
 public record ReservationTimeRequest(@NotNull LocalTime startAt) {
     public ReservationTime toReservationTime() {
