+++ src/main/java/roomescape/application/dto/ReservationTimeResponse.java
@@ -1,7 +1,7 @@
 package roomescape.application.dto;
 
 import java.time.LocalTime;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservation.ReservationTime;
 
 public record ReservationTimeResponse(long id, LocalTime startAt) {
     public static ReservationTimeResponse from(ReservationTime reservationTime) {
