+++ src/main/java/roomescape/application/dto/ThemeResponse.java
@@ -1,6 +1,6 @@
 package roomescape.application.dto;
 
-import roomescape.domain.Theme;
+import roomescape.domain.theme.Theme;
 
 public record ThemeResponse(long id, String name, String description, String thumbnail) {
     public static ThemeResponse from(Theme theme) {
