+++ src/main/java/roomescape/domain/reservation/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,6 +7,8 @@ import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import java.time.LocalDate;
 import java.util.Objects;
+import roomescape.domain.member.Member;
+import roomescape.domain.theme.Theme;
 import roomescape.exception.RoomescapeErrorCode;
 import roomescape.exception.RoomescapeException;
