+++ src/main/java/roomescape/domain/reservation/ReservationCommandRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
