+++ src/main/java/roomescape/domain/reservation/ReservationFactory.java
@@ -1,8 +1,13 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.time.Clock;
 import java.time.LocalDateTime;
 import roomescape.application.dto.ReservationRequest;
+import roomescape.domain.DomainService;
+import roomescape.domain.member.Member;
+import roomescape.domain.member.MemberQueryRepository;
+import roomescape.domain.theme.Theme;
+import roomescape.domain.theme.ThemeRepository;
 import roomescape.exception.RoomescapeErrorCode;
 import roomescape.exception.RoomescapeException;
