+++ src/main/java/roomescape/domain/reservation/ReservationQueryRepository.java
@@ -1,10 +1,11 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import roomescape.domain.dto.AvailableTimeDto;
+import roomescape.domain.theme.Theme;
 
 public interface ReservationQueryRepository extends JpaRepository<Reservation, Long> {
     long countByTimeId(long timeId);
