+++ src/test/java/roomescape/domain/ThemeNameTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.domain.theme.ThemeName;
 import roomescape.exception.RoomescapeErrorCode;
 import roomescape.exception.RoomescapeException;
