+++ src/test/java/roomescape/domain/member/EmailTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.member;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.domain.member.Email;
 import roomescape.exception.RoomescapeErrorCode;
 import roomescape.exception.RoomescapeException;
