+++ src/test/java/roomescape/domain/member/PasswordTest.java
@@ -1,10 +1,11 @@
-package roomescape.domain;
+package roomescape.domain.member;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.domain.member.Password;
 import roomescape.exception.RoomescapeErrorCode;
 import roomescape.exception.RoomescapeException;
